requirements.txt
SQLAlchemy==1.3.20
requests==2.21.0
uvicorn==0.13.0
nfcpy==1.0.3
starlette==0.13.6
fastapi==0.62.0
nfc==0.0.1
PyYAML==5.3.1